Uploaded dispatcher.v
The Dispatcher coordinates the distribution of work (blocks of threads) across our 2 cores. It uses the thread_count from the DCR to determine how many threads to process and assigns blocks of 4 threads to each core. It ensures all threads in a kernel are executed and signals when the kernel is complete.